+++ src/core/object/watch/engines/accessors.ts
@@ -107,7 +107,7 @@ export function watch<T extends object>(
 
 			set: (path, value) => {
 				if (handlers) {
-					set(obj, path, value, handlers);
+					set.call(opts?.engine)(obj, path, value, handlers);
 				}
 			},
 
@@ -179,7 +179,7 @@ export function watch<T extends object>(
 			unwrappedObj,
 			toProxyObject,
 			handlers,
-			() => unwrappedObj.slice()
+			unwrappedObj
 		);
 
 		for (let i = 0; i < (<unknown[]>proxy).length; i++) {